+++ src/render.py
@@ -298,8 +298,8 @@ def ApplyRotation(img, rot):
 # generate a dummy image
 def DummyPage():
     img = Image.new('RGB', (ScreenWidth, ScreenHeight))
-    img.paste(LogoImage, ((ScreenWidth  - LogoImage.size[0]) / 2,
-                          (ScreenHeight - LogoImage.size[1]) / 2))
+    img.paste(LogoImage, (int((ScreenWidth  - LogoImage.size[0]) / 2),
+                          int((ScreenHeight - LogoImage.size[1]) / 2)))
     return img
 
 # load a page from a PDF file
@@ -560,12 +560,12 @@ def PageImage(page, ZoomMode=False, RenderMode=False):
         # create black background image to paste real image onto
         if ZoomMode:
             TextureImage = Image.new('RGB', (int(ResZoomFactor * TexWidth), int(ResZoomFactor * TexHeight)))
-            TextureImage.paste(img, ((int(ResZoomFactor * ScreenWidth)  - img.size[0]) / 2, \
-                                     (int(ResZoomFactor * ScreenHeight) - img.size[1]) / 2))
+            TextureImage.paste(img, (int((ResZoomFactor * ScreenWidth  - img.size[0]) / 2), \
+                                     int((ResZoomFactor * ScreenHeight - img.size[1]) / 2)))
         else:
             TextureImage = Image.new('RGB', (TexWidth, TexHeight))
-            x0 = (ScreenWidth  - img.size[0]) / 2
-            y0 = (ScreenHeight - img.size[1]) / 2
+            x0 = int((ScreenWidth  - img.size[0]) / 2)
+            y0 = int((ScreenHeight - img.size[1]) / 2)
             TextureImage.paste(img, (x0, y0))
             SetPageProp(page, '_box', (x0, y0, x0 + img.size[0], y0 + img.size[1]))
             FixHyperlinks(page)
@@ -582,12 +582,12 @@ def PageImage(page, ZoomMode=False, RenderMode=False):
                 # first, fill the underlying area with black (i.e. remove the dummy logo)
                 blackness = Image.new('RGB', (OverviewCellX - OverviewBorder, \
                                               OverviewCellY - OverviewBorder))
-                OverviewImage.paste(blackness, (pos[0] + OverviewBorder / 2, \
+                OverviewImage.paste(blackness, (pos[0] + int(OverviewBorder / 2), \
                                                 pos[1] + OverviewBorder))
                 del blackness
                 # then, scale down the original image and paste it
                 if HalfScreen:
-                    img = img.crop((0, 0, img.size[0] / 2, img.size[1]))
+                    img = img.crop((0, 0, int(img.size[0] / 2), img.size[1]))
                 sx = OverviewCellX - 2 * OverviewBorder
                 sy = OverviewCellY - 2 * OverviewBorder
                 if HighQualityOverview:
@@ -601,8 +601,8 @@ def PageImage(page, ZoomMode=False, RenderMode=False):
                     img.thumbnail((sx * 2, sy * 2), Image.NEAREST)
                     img.thumbnail((sx, sy), Image.BILINEAR)
                 OverviewImage.paste(img, \
-                   (pos[0] + (OverviewCellX - img.size[0]) / 2, \
-                    pos[1] + (OverviewCellY - img.size[1]) / 2))
+                   (pos[0] + int((OverviewCellX - img.size[0]) / 2), \
+                    pos[1] + int((OverviewCellY - img.size[1]) / 2)))
             finally:
                 Loverview.release()
             SetPageProp(page, '_overview_rendered', True)
